+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/emulators/EmulatorChangeListener.java
@@ -1,0 +1,5 @@
+package de.mephisto.vpin.server.emulators;
+
+public interface EmulatorChangeListener {
+  void emulatorChanged(int emulatorId);
+}