+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/emulators/EmulatorService.java
@@ -3,10 +3,10 @@ package de.mephisto.vpin.server.emulators;
 import de.mephisto.vpin.restclient.validation.ValidationState;
 import de.mephisto.vpin.server.frontend.FrontendConnector;
 import de.mephisto.vpin.server.frontend.FrontendService;
+import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.games.GameEmulator;
 import de.mephisto.vpin.server.games.GameEmulatorValidationService;
 import de.mephisto.vpin.server.mame.MameService;
-
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,6 +26,8 @@ public class EmulatorService {
 
   @Autowired
   private MameService mameService;
+
+  private List<EmulatorChangeListener> listeners = new ArrayList<>();
 
   private final Map<Integer, GameEmulator> emulators = new LinkedHashMap<>();
 
@@ -84,58 +86,6 @@ public class EmulatorService {
     return getGameEmulators().stream().filter(GameEmulator::isValid).filter(e -> e.isVpxEmulator()).collect(Collectors.toList());
   }
 
-  //@deprecated, always determine emulator
-//  public GameEmulator getDefaultGameEmulator() {
-//    Collection<GameEmulator> emulators = this.emulators.values();
-//
-//    // when there is only one VPX emulator, it is forcibly the default one
-//    if (emulators.size() == 1) {
-//      GameEmulator value = emulators.iterator().next();
-//      return value.isVpxEmulator() ? value : null;
-//    }
-//
-//    for (GameEmulator emulator : emulators) {
-//      if (emulator.isValid() && emulator.getDescription() != null && emulator.isVpxEmulator() && emulator.getDescription().contains("default")) {
-//        return emulator;
-//      }
-//    }
-//
-//    for (GameEmulator value : emulators) {
-//      if (value.isValid() && value.isVpxEmulator() && value.getNvramFolder().exists()) {
-//        return value;
-//      }
-//      else {
-//        // avoid NPE when installationFolder is null like in test
-//        if (value.isValid()) {
-//          LOG.error(value + " has no nvram folder \"" + value.getNvramFolder().getAbsolutePath() + "\"");
-//        }
-//        else {
-//          LOG.error(value + " has no valid nvram folder");
-//        }
-//      }
-//    }
-//    LOG.error("Failed to determine emulator for highscores, no VPinMAME/nvram folder could be resolved (" + this.emulators.size() + " VPX emulators found).");
-//    return null;
-//  }
-
-  public boolean isValidVPXEmulator(GameEmulator emulator) {
-    if (!emulator.getType().isVpxEmulator()) {
-      return false;
-    }
-
-    if (StringUtils.isEmpty(emulator.getGamesDirectory())) {
-      LOG.warn("Ignoring " + emulator + ", because \"Games Folder\" is not set.");
-      return false;
-    }
-
-    if (frontendService.getFrontendConnector().getMediaAccessStrategy() != null && StringUtils.isEmpty(emulator.getMediaDirectory())) {
-      LOG.warn("Ignoring " + emulator + ", because \"Media Dir\" is not set.");
-      return false;
-    }
-
-    return true;
-  }
-
   public void setFrontendService(FrontendService frontendService) {
     this.frontendService = frontendService;
   }
@@ -144,12 +94,15 @@ public class EmulatorService {
     GameEmulator saved = frontendService.saveEmulator(emulator);
     this.emulators.remove(saved.getId());
     loadEmulator(saved);
+    synchronizeEmulator(emulator);
+    notifyEmulatorChange(emulator.getId());
     return saved;
   }
 
   public boolean delete(int emulatorId) {
     frontendService.deleteEmulator(emulatorId);
     this.emulators.remove(emulatorId);
+    notifyEmulatorChange(emulatorId);
     return true;
   }
 
@@ -162,6 +115,10 @@ public class EmulatorService {
       loadEmulator(emulator);
     }
 
+    for (GameEmulator emulator : ems) {
+      synchronizeEmulator(emulator);
+    }
+
     if (this.emulators.isEmpty()) {
       LOG.error("*****************************************************************************************");
       LOG.error("No valid game emulators folder, fill all(!) emulator directory settings in your frontend.");
@@ -171,11 +128,6 @@ public class EmulatorService {
 
   private void loadEmulator(GameEmulator emulator) {
     try {
-//      if (emulator.getType().isVpxEmulator() && !isValidVPXEmulator(emulator)) {
-//        return;
-//      }
-
-
       File mameFolder = new File(emulator.getInstallationDirectory(), "VPinMAME");
       if (mameFolder.exists()) {
         emulator.setMameDirectory(mameFolder.getAbsolutePath());
@@ -188,7 +140,7 @@ public class EmulatorService {
         }
         else {
           emulator.setNvramDirectory(new File(mameFolder, "nvram").getAbsolutePath());
-        }  
+        }
 
         File cfgFolder = mameService.getCfgFolder();
         if (cfgFolder != null && cfgFolder.exists()) {
@@ -202,7 +154,7 @@ public class EmulatorService {
         if (StringUtils.isEmpty(emulator.getRomDirectory())) {
           File romFolder = mameService.getRomsFolder();
           if (romFolder != null && romFolder.exists()) {
-            emulator.setRomDirectory(romFolder.getAbsolutePath());;
+            emulator.setRomDirectory(romFolder.getAbsolutePath());
           }
           else {
             emulator.setRomDirectory(new File(mameFolder, "roms").getAbsolutePath());
@@ -218,8 +170,52 @@ public class EmulatorService {
     }
   }
 
+  private void synchronizeEmulator(GameEmulator emulator) {
+    if (!emulator.isEnabled()) {
+      return;
+    }
+
+    if (emulator.isFxEmulator()) {
+      synchronizeEmulatorGames(emulator);
+    }
+  }
+
+  private void synchronizeEmulatorGames(GameEmulator emulator) {
+    int count = 0;
+    List<Game> gamesByEmulator = frontendService.getGamesByEmulator(emulator.getId());
+    String gamesDirectory = emulator.getGamesDirectory();
+    if (gamesDirectory != null) {
+      File gamesFolder = new File(gamesDirectory);
+      if (gamesFolder.exists()) {
+        File[] files = gamesFolder.listFiles((dir, name) -> name.endsWith(emulator.getGameExt()));
+        if (files != null) {
+          for (File file : files) {
+            Optional<Game> game = gamesByEmulator.stream().filter(g -> g.getGameFile().equals(file)).findFirst();
+            if (game.isEmpty()) {
+              LOG.info("Importing \"{}\" for emulator \"{}\".", file.getAbsolutePath(), emulator.getName() + "/" + emulator.getId());
+              frontendService.importGame(file, emulator.getId());
+              count++;
+            }
+          }
+        }
+      }
+    }
+
+    LOG.info("\"{}\" emulator synchronization finished, added {} games.", emulator.getName(), count);
+  }
+
   public boolean clearCache() {
     loadEmulators();
     return true;
+  }
+
+  private void notifyEmulatorChange(int emulatorId) {
+    for (EmulatorChangeListener listener : listeners) {
+      listener.emulatorChanged(emulatorId);
+    }
+  }
+
+  public void addEmulatorChangeListener(EmulatorChangeListener listener) {
+    this.listeners.add(listener);
   }
 }