+++ RELEASE_NOTES.md
@@ -15,7 +15,7 @@
 - **VPS Overview**:
   - **Tutorial Column**: Improved the "Tutorials" column. The column shows the available tutorials for the given table. Note that the **Kongedam** tutorials have an additional color indicator. (This indicator has a technical reason too. It ensures that all the latest and greatest from him are available in the pause menu too). 
 - **Table Overview**: 
-  - In Table Data dialog, control the overriden rom Name exists
+  - In Table Data dialog, control the overridden rom Name exists
   - **ALT Color Backups**: Added dialog for backups to the ALT color section. Note that you may already have some backups available since the actual backup support for the backend was already build a year ago. The dialog allows you to restore these files now.
 - **Backglass Manager**:
   - Improved refresh of backglass sidebar when multiple clicks in list
@@ -24,7 +24,12 @@
 - **Main Window**:
   - Added options to start and stop PinVOL directly from the preferences dropdown menu. Since PinVOL may block key bindings, you can faster work around this problem this way.
 - **Uploads**:
-  - Improved ROM name detection of ALT sound bundles.
+  - Fixed and improved ROM name detection of ALT sound bundles.
+- **Emulator Management**: 
+  - Added missing game deletion from the database when an emulator is deleted.
+  - Added missing server-side cache invalidations after emulator changes.
+  - Added missing game scan for FX emulators so that these are automatically imported when the emulator is set up via Studio. Note that you have to set the proper game folder and file extension for this.
+  - The "Import Game" dialog is now enabled for all emulators. After creating a new emulator the dialog can be used to import the related games. This was limited to FP and VPX before.
 
 
 ## VPin Mania